logger: simplified
